% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my2ddnorm.R
\name{my2ddnorm}
\alias{my2ddnorm}
\title{2D, uncorrelated, equal variance, Gaussian density function}
\usage{
my2ddnorm(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{mean}{vector of means.}

\item{sd}{vector of standard deviations.}
}
\value{
A vector of density values.
}
\description{
Function for calculating the density of a uncorrelated 2 dimensional normal distribution with equal variance.
}
