% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGaussian_child.R
\name{rGaussian_child}
\alias{rGaussian_child}
\title{Generate clusters disperced with Gaussian density}
\usage{
rGaussian_child(parents, sd, mean, var = NULL, ncores = 1)
}
\arguments{
\item{parents}{An object of class \code{\link{ppp}}. Determining the possition of the centre points}

\item{sd}{A single non-negative numeric or a vector consisting of two non-negative numerics.
May also be a matrix such that squaring each entry yields the covariance matrix.}

\item{mean}{A non-negative numeric. The mean number of points in a single cluster.}

\item{var}{A non-negative numeric. The variance of the number of points in a cluster. See details.}

\item{ncores}{Positive integer specifying the number of cores to be used.}
}
\value{
A point pattern of class \code{\link{ppp}} in the same window as \code{parents}.
}
\description{
Generate clusters disperced by a normal distribution around a center point.
}
\details{
Number of point are Poisson distributed if \code{var} is \code{NULL} or identical to \code{mean}.
It is negativel binomially distributed if \code{var>mean} and binomially distributed if \code{var<mean}.
}
