% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricpp_sr.R
\name{ricpp_sr}
\alias{ricpp_sr}
\title{Simulate iterated cluster point prcocesses with independent imigration}
\usage{
ricpp_sr(intens, clsiz, sd, noise, win = owin(), kernel_sd = NULL,
  clsiz_var = NULL, n = NULL, tol = .Machine$double.eps,
  ncores = 1, track = FALSE)
}
\arguments{
\item{intens}{Non-negative numeric. The desired intensity for each generation.}

\item{clsiz}{Numeric between 0 and 1. The expected cluster size.
Must be between 0 and 1, due to the assumption of same reproduction.}

\item{sd}{Non-negative numeric. The offspring density standard deviation.}

\item{noise}{String describing the type of noise to be added. \code{"Det"}, \code{"Pois"}, or \code{"Per"}.}

\item{win}{Window of class \code{\link{owin}}.}

\item{kernel_sd}{Non-negative numeric. Kernel standard deviation.
Default is \code{NULL} in which case \code{noise} must be \code{"Pois"}.}

\item{clsiz_var}{Non-negative numeric. Variance of the number of points in a cluster. See \code{\link{rGaussian_child}}.}

\item{n}{Positive integer. The number of iterations.}

\item{tol}{Positive real number.}

\item{ncores}{Positive integer specifying the number of cores to be used.}

\item{track}{logical indicating wheter to print the iterationnumber for each completed iteration.
Is not necessarily in chronological orden when \code{ncores > 1}.
See the argument \code{mc.cores} in \code{\link{mclapply}}.}
}
\description{
Simulate iterated cluster point prcocesses with independent imigration under same reproduction.
}
