% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPPP.R
\name{rPPP}
\alias{rPPP}
\title{Simulate a weighted-permanental point pattern}
\usage{
rPPP(intens, kernel_sd, alpha = 0.5, win = owin())
}
\arguments{
\item{intens}{Non-negative numeric. Desired intensity.}

\item{kernel_sd}{Non-negative numeric. Kernel standard deviation.}

\item{alpha}{Half integer. Default is \code{1/2}. The weight of the weighted permanental point process.}

\item{win}{Window of class \code{\link{owin}}. Default is \code{owin()}.}
}
\value{
A pixel image of class \code{\link{im}}.
}
\description{
Simulate an alpha-weighted-permanental point pattern when alpha is half an integer. This may be regarded as a Cox process.
}
